topics with @host, nut physics to int32

+++ src/agents/nut/nut-driver.cc
@@ -384,8 +384,8 @@ std::map<std::string,std::string> NUTDevice::properties() const {
     return map;
 }
 
-std::map<std::string,int64_t> NUTDevice::physics(bool onlyChanged) const {
-    std::map<std::string,int64_t> map;
+std::map<std::string,int32_t> NUTDevice::physics(bool onlyChanged) const {
+    std::map<std::string,int32_t> map;
     for(auto it : _physics ){
         if( ( ! onlyChanged ) || it.second.changed ) {
             map[ it.first ] = it.second.value;
